requirements.txt
pygame==2.5.2
neat-python==0.92
numpy>=1.20.0
Pillow>=9.0.0
